requirements.txt
discord.py>=2.0.0
requests>=2.28.0
pytz>=2022.1
python-dotenv>=0.20.0
beautifulsoup4>=4.11.0
aiohttp>=3.8.1
asyncio>=3.4.3
datetime>=4.4